Final project
Objective: You will design and implement a text-based game or puzzle where the player moves
through a series of rooms or compartments. Each space will be a class with (at least) four pointer
variables that link to other spaces. You must have at least 5 spaces of at least 3 different types. You will have a space abstract class that will have a special pure virtual function. Each type of space will have a special action. You will have at least 3 derived classes for different types of spaces. You will have at least 3 derived classes for different types of spaces.
You must have some way to keep track of which space the player is in. The player will have a container (backpack, knitting bag, or notebook) to carry “items”. The container must have some limit. One or more of these items will be required as part of the solution, such as a “key” to open the locked door. You should have a time limit to urge the player on. This does not mean a literal clock, just some way to prevent the ‘game’ from going on indefinitely. The player must interact with parts of the structure, and not just simply collect things.This can be throwing something at the monster, operating a light switch (or other control), opening doors, or singing to get the baby back to sleep.